add react-query signup logic
