+++ src/presentation/components/organisms/SignUpForm.tsx
@@ -1,11 +1,7 @@
 import { TextField, Container, FormControl, Button } from "@mui/material";
-import { signUp } from "../../../services/authHttpClient.adapter";
-import { useAppDispatch } from "../../hooks/reduxHooks";
-import { authActions } from "../../../services/redux/authSlice";
-import { saveToken, saveUserId } from "../../../services/localStorage.adapter";
-import { useNavigate } from "react-router-dom";
-import useForm from "../../hooks/useForm";
 import useErrorHandler from "../../hooks/useErrorHandler";
+import { useAuthMutations } from "../../hooks/useAuthMutations";
+import useForm from "../../hooks/useForm";
 
 const initialValues = {
   firstName: "test first name",
@@ -17,34 +13,30 @@ const initialValues = {
 const SignUpForm = () => {
   const { values, valueChangeHandler } = useForm(initialValues);
   const { firstName, lastName, email, password } = values;
-  const dispatch = useAppDispatch();
-  const navigate = useNavigate();
   const { handleError } = useErrorHandler();
+  const { signUpMutation } = useAuthMutations();
 
   const submitHandler: React.FormEventHandler<HTMLFormElement> = async (
     event
   ) => {
     event.preventDefault();
 
-    try {
-      const { token, user } = await signUp({
-        firstName,
-        lastName,
-        email,
-        password,
-      });
-      const { _id: userId } = user;
-
-      dispatch(authActions.signUp({ token, userId }));
-
-      saveToken(token);
-      saveUserId(userId);
-
-      navigate("/");
-    } catch (error) {
-      handleError(error);
-    }
+    signUpMutation.mutate({
+      firstName,
+      lastName,
+      email,
+      password,
+    });
   };
+
+  if (signUpMutation.isLoading) {
+    return <div>Signing up...</div>;
+  }
+
+  if (signUpMutation.isError) {
+    handleError(signUpMutation.error);
+    signUpMutation.reset();
+  }
 
   return (
     <Container>
