+++ src/presentation/pages/SignUp.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import SignUpForm from "../components/organisms/SignUpForm";
 
 const SignUp = () => {
